Start shoving readline logic into readline.c
