+++ prim.c
@@ -48,6 +48,10 @@ extern void initprims(void) {
 	prims = initprims_proc(prims);
 	prims = initprims_access(prims);
 
+#if HAVE_READLINE
+	prims = initprims_readline(prims);
+#endif
+
 #define	primdict prims
 	X(primitives);
 }
