+++ prim.h
@@ -19,3 +19,7 @@ extern Dict *initprims_etc(Dict *primdict);		/* prim-etc.c */
 extern Dict *initprims_sys(Dict *primdict);		/* prim-sys.c */
 extern Dict *initprims_proc(Dict *primdict);		/* proc.c */
 extern Dict *initprims_access(Dict *primdict);		/* access.c */
+
+#if HAVE_READLINE
+extern Dict *initprims_readline(Dict *primdict);	/* readline.c */
+#endif
